A dynamic set of tabs, at least with regards to Version Control API's potentially large number VCS accounts, is near to impossible with Drupal 6. (Sam knows it, he practically wrote the book on that topic.)
Therefore, fix the remaining menu system hassles by throwing it all
into a single "Repository accounts" tab (in the user edit section),
and filling it with either a VCS account edit form or a list of accounts
that forwards to that VCS account edit form. Also, the insane
hook_user() hack has made a disappearance due to that new method.

Works reasonably well and comes with the advantage that it will now even
look exactly like on drupal.org ("CVS" tab title if appropriate), the only
issue that I couldn't find a solution for is that the second row of tabs
is lost when additional query arguments are passed. Bug in D6? hm.